requirements.txt
streamlit
python-dotenv
pdf2image
google-generativeai
PyPDF2
pdfplumber
python-docx
spacy
keybert
textstat
rapidfuzz
llama-cpp-python
matplotlib
pandas
argparse
rich
ollama
textdescriptives
plotly
groq
flask
flask_sqlalchemy
streamlit-aggrid